refactor(viewer): share list group-sort from lib, fix tie-break direction
Address review feedback on #1540:

- Move the pure cell comparator + group-header ordering into
  lib/lists/group-sort.ts so both the table component and the export
  model depend downward on lib, instead of the export model (in lib/)
  importing from a components/ util (layer inversion). list-table-utils
  re-exports them so existing consumers are untouched.
- Fold the label tie-breaker into the direction multiplier so groups
  that tie on a summed column resolve their labels in the sort
  direction, not always ascending.
