+++ apps/viewer/src/components/viewer/lists/list-table-utils.ts
@@ -9,6 +9,11 @@
  */
 
 import type { CellValue, ColumnDefinition, ListRow, ListGrouping } from '@ifc-lite/lists';
+import { compareCells, orderGroups, type GroupSort, type OrderableGroup } from '@/lib/lists/group-sort';
+
+// Re-exported so existing consumers keep importing the list-table barrel.
+export { compareCells, orderGroups };
+export type { GroupSort, OrderableGroup };
 
 export function formatCellValue(value: CellValue): string {
   if (value === null || value === undefined) return '';
@@ -18,14 +23,6 @@ export function formatCellValue(value: CellValue): string {
     return value.toFixed(4).replace(/\.?0+$/, '');
   }
   return String(value);
-}
-
-export function compareCells(a: CellValue, b: CellValue): number {
-  if (a === null && b === null) return 0;
-  if (a === null) return -1;
-  if (b === null) return 1;
-  if (typeof a === 'number' && typeof b === 'number') return a - b;
-  return String(a).localeCompare(String(b));
 }
 
 /** A column is numeric (summable) when every sampled non-empty value is a
@@ -65,49 +62,13 @@ export type DisplayItem =
 export interface Totals { count: number; sums: Record<string, number> }
 export interface GroupedView { items: DisplayItem[]; groupCount: number; totals: Totals }
 
-/** Active header sort — a column index into the result columns and a
- *  direction. `null` means no explicit sort (grouped view falls back to its
- *  count-descending default). */
-export type GroupSort = { colIdx: number; dir: 'asc' | 'desc' } | null;
-
 function sumIndices(columns: ColumnDefinition[], sumColumnIds: string[]) {
   return sumColumnIds
     .map((id) => ({ id, idx: columns.findIndex((c) => c.id === id) }))
     .filter((s) => s.idx >= 0);
 }
 
-interface GroupBucket { key: string; label: string; raw: CellValue; count: number; sums: Record<string, number>; rows: ListRow[] }
-
-/** Minimal shape `orderGroups` needs — the group header's raw value (for the
- *  group-by column), row count, and per-column subtotals. */
-export interface OrderableGroup { label: string; raw: CellValue; count: number; sums: Record<string, number> }
-
-/** Order the group headers to match the active header sort so that toggling
- *  a column's sort actually reorders the groups (not just the rows inside
- *  them). The group header only carries a value for the group-by column and
- *  the summed columns, so those drive the order; anything else (or no sort)
- *  falls back to the count-descending default. Shared with the export path so
- *  the exported sections keep the on-screen order. */
-export function orderGroups<T extends OrderableGroup>(
-  groups: T[],
-  sort: GroupSort,
-  groupIdx: number,
-  sums: { id: string; idx: number }[],
-): T[] {
-  const byCount = (a: T, b: T) => b.count - a.count || a.label.localeCompare(b.label);
-  if (!sort) return groups.sort(byCount);
-
-  const dir = sort.dir === 'desc' ? -1 : 1;
-  if (sort.colIdx === groupIdx) {
-    return groups.sort((a, b) => compareCells(a.raw, b.raw) * dir || a.label.localeCompare(b.label));
-  }
-  const summed = sums.find((s) => s.idx === sort.colIdx);
-  if (summed) {
-    return groups.sort((a, b) => (a.sums[summed.id] - b.sums[summed.id]) * dir || a.label.localeCompare(b.label));
-  }
-  // A non-group, non-summed column has no group-level value to order by.
-  return groups.sort(byCount);
-}
+interface GroupBucket extends OrderableGroup { key: string; rows: ListRow[] }
 
 /** Bucket already-filtered/sorted rows by the group-by column, accumulate
  *  per-group + grand count/sums, and flatten into a virtualizable list
